requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.8
aiosignal==1.3.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
dataclasses-json==0.6.7
Deprecated==1.2.15
distro==1.9.0
duckduckgo_search==6.3.7
faiss-cpu==1.9.0.post1
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.10.0
g4f==0.3.7.6
gitdb==4.0.11
GitPython==3.1.43
googlesearch-python==1.2.3
greenlet==3.1.1
groq==0.12.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
httpx-sse==0.4.0
huggingface-embeddings==0.1.1
huggingface-hub==0.26.3
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.9
langchain-community==0.3.8
langchain-core==0.3.21
langchain-document-loaders==0.0.27
langchain-groq==0.2.1
langchain-text-splitters==0.3.2
langsmith==0.1.147
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.14.2
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
orjson==3.10.12
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
plotly==5.24.1
primp==0.8.1
propcache==0.2.0
protobuf==5.29.0
pyarrow==18.1.0
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.2
pydantic-settings==2.6.1
pydantic-core==2.27.1
pydeck==0.9.1
PyGithub==2.5.0
Pygments==2.18.0
PyJWT==2.10.1
PyNaCl==1.5.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.21.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
selenium==4.27.1
sentence-transformers==3.3.1
setuptools==75.6.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.35
streamlit==1.32.0
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.20.3
toml==0.10.2
torch==2.5.1
tornado==6.4.2
tqdm==4.67.1
transformers==4.46.3
trio==0.27.0
trio-websocket==0.11.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==6.0.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wrapt==1.17.0
wsproto==1.2.0
yarl==1.18.0
